feat: unify file naming convention across project
- Renamed files to follow a consistent convention.
- Improved maintainability and readability by standardizing file structure.
