+++ src/types/index.ts
@@ -1,0 +1,28 @@
+import { ErrorStatusCode } from "@/types/status-code.type";
+import { Meta } from "./response.type";
+
+/* eslint-disable @typescript-eslint/no-namespace */
+export type MyEnvs = {
+  PORT: string;
+  NODE_ENV: "development" | "production";
+  BASE_URL: string;
+  APP_DEBUG: string;
+};
+
+declare global {
+  namespace NodeJS {
+    interface ProcessEnv extends MyEnvs {}
+  }
+
+  namespace Express {
+    interface Response {
+      sendSuccess<T>(data: T, message?: string, meta?: Meta): void;
+      sendError(
+        message: string,
+        statusCode?: ErrorStatusCode,
+        errors?: Record<string, string>,
+      ): void;
+      sendCreated<T>(data: T, message?: string, meta?: Meta): void;
+    }
+  }
+}
